Benchmark done

+++ Person.jl
@@ -1,7 +1,7 @@
 struct Person 
     start_point::Int
-    start_time::Int
-    end_time::Int
+    start_time::Float64
+    end_time::Float64
 
     Person(; start_point, start_time, end_time) = new(start_point, start_time, end_time)
 end
